:recycle: api: move usecase into sessions-mass-import directory
