+++ api/lib/domain/usecases/sessions-mass-import/validate-sessions.js
@@ -1,11 +1,11 @@
-const Session = require('../models/Session');
-const sessionCodeService = require('../services/session-code-service');
-const sessionsImportValidationService = require('../services/sessions-mass-import/sessions-import-validation-service');
-const temporarySessionsStorageForMassImportService = require('../services/sessions-mass-import/temporary-sessions-storage-for-mass-import-service');
-const CertificationCandidate = require('../models/CertificationCandidate');
+const Session = require('../../models/Session');
+const sessionCodeService = require('../../services/session-code-service');
+const sessionsImportValidationService = require('../../services/sessions-mass-import/sessions-import-validation-service');
+const temporarySessionsStorageForMassImportService = require('../../services/sessions-mass-import/temporary-sessions-storage-for-mass-import-service');
+const CertificationCandidate = require('../../models/CertificationCandidate');
 const bluebird = require('bluebird');
 
-module.exports = async function validateSessionsForMassImport({
+module.exports = async function validateSessions({
   sessions,
   userId,
   certificationCenterId,
